requirements.txt
langchain
openai
wikipedia
python-dotenv
langchain-community
duckduckgo-search
langchain-openai